Filter physnets when ports are pre-allocated to a segment
If an Ironic Node is connected to multiple physnets, and wants to
connect to a Neutron VIF whose network has segments in multiple
physnets it may be unclear which one to connect to.

If the VIF is already allocated to a subnet, this can be used to
resolve the process to a single segment and hence physnet.

When a network uses Neutron routed segments, this patch first
checks the port's fixed_ips to see if a single physnet can be
determined from the segment it is attached to. If this cannot be
identified (or fixed_ips is unset), the behaviour falls back to
returning all physnets available in the segmented network.